lazy.nvim migration + zed
