--- nvim/after/plugin/multiple.lua
@@ -1,3 +1,0 @@
-vim.g.VM_leader = '<C+/>'
-vim.g.VM_maps["Add Cursor Down"] = '<S-C-Down>'
-vim.g.VM_maps["Add Cursor Up"] = '<S-C-Up>'
