--- nvim/after/plugin/which_key.lua
@@ -1,10 +1,0 @@
-vim.o.timeout = true
-vim.o.timeoutlen = 300
-require("which-key").setup {
-    layout = {
-        height = { min = 4, max = 60 }, -- min and max height of the columns
-        width = { min = 20, max = 50 }, -- min and max width of the columns
-        spacing = 3,                    -- spacing between columns
-        aligno = "left",                 -- aligno columns left, center or right
-    },
-}
