+++ nvim/lua/nogod/init.lua
@@ -1,3 +1,4 @@
+-- require('nogod.packer')
 require("nogod.remap")
-require('nogod.packer')
 require('nogod.set')
+require("nogod.lazy")
