+++ nvim/lua/nogod/plugins/colors.lua
@@ -5,4 +5,14 @@ function ColorMyPencils(color)
 	vim.api.nvim_set_hl(0, "Normal", { bg = "none"}) 
 	vim.api.nvim_set_hl(0, "NormalFloat", { bg = "none"}) 
 end
-ColorMyPencils()
+return {
+    'rose-pine/neovim',
+    name = 'rose-pine',
+    config = function()
+        vim.cmd('colorscheme rose-pine')
+        ColorMyPencils()
+    end,
+    lazy = false,
+    priority = 150,
+
+}
