+++ nvim/lua/nogod/plugins/fugitive.lua
@@ -1,0 +1,6 @@
+return {
+    'tpope/vim-fugitive',
+    config = function()
+        vim.keymap.set("n", "<leader>gg", vim.cmd.Git);
+    end,
+}
