+++ nvim/lua/nogod/plugins/harpoon.lua
@@ -1,0 +1,16 @@
+return {
+    'ThePrimeagen/harpoon',
+    config = function()
+        local mark = require("harpoon.mark")
+        local ui = require("harpoon.ui")
+
+
+        vim.keymap.set("n", "<leader>a", mark.add_file)
+        vim.keymap.set("n", "<C-e>", ui.toggle_quick_menu)
+
+        vim.keymap.set("n", "<leader>1",function() ui.nav_file(1) end)
+        vim.keymap.set("n", "<leader>2",function() ui.nav_file(2) end)
+        vim.keymap.set("n", "<leader>3",function() ui.nav_file(3) end)
+        vim.keymap.set("n", "<leader>4",function() ui.nav_file(4) end)
+    end,
+}
