+++ nvim/lua/nogod/plugins/mason.lua
@@ -1,0 +1,7 @@
+return {
+    'williamboman/mason.nvim',
+    name = 'mason',
+    config = function()
+        require('mason').setup({})
+    end,
+}
