+++ nvim/lua/nogod/plugins/plantuml.lua
@@ -1,0 +1,6 @@
+return {
+  'https://gitlab.com/itaranto/plantuml.nvim',
+  version = '*',
+  config = function() require('plantuml').setup() end,
+}
+
