+++ nvim/lua/nogod/remap.lua
@@ -47,3 +47,68 @@ vim.keymap.set("n", "<leader>mr", "<cmd>CellularAutomaton make_it_rain<CR>");
 vim.keymap.set("n", "<leader><leader>", function()
     vim.cmd("so")
 end)
+
+local function open_split_buffer_goto_definition()
+    -- Check how many windows are open
+    local wins = vim.api.nvim_tabpage_list_wins(0)
+    local current_cursor_pos = vim.api.nvim_win_get_cursor(0)
+
+    local current_buf = vim.api.nvim_get_current_buf()
+    local current_win = vim.api.nvim_get_current_win()
+
+    local target_win
+
+    -- If there are already splits, then take the next one and set buffer to current buffer
+    if #wins >= 2 then
+        target_win = current_win
+        for _, win_num in pairs(wins) do
+
+            if win_num < target_win then
+                target_win = win_num
+            end
+        end
+    else
+        target_win = current_win
+        vim.cmd("vsplit")
+    end
+
+    -- Set buffer for new window
+    vim.api.nvim_win_set_buf(target_win, current_buf)
+
+    -- Copy cursor position to new window for lsp defintion
+    vim.api.nvim_win_set_cursor(target_win, current_cursor_pos)
+
+    -- Focus new window
+    vim.api.nvim_set_current_win(target_win)
+
+    -- Call lsp
+    vim.lsp.buf.definition()
+end
+
+local augroup = vim.api.nvim_create_augroup   -- Create/get autocommand group
+local autocmd = vim.api.nvim_create_autocmd
+nogod = augroup("nogod", {})
+autocmd('LspAttach', { 
+
+	group = nogod,
+	callback = function(e)
+
+		local opts = {buffer = e.buf, remap = false}
+
+        vim.keymap.set("n", "<S-F6>", function()
+            vim.lsp.buf.rename()
+        end)
+		-- vim.keymap.set("n", "gd", function() vim.lsp.buf.definition() end, opts)
+		vim.keymap.set("n", "gd", open_split_buffer_goto_definition, {})
+		vim.keymap.set("n", "K", function() vim.lsp.buf.hover() end, opts)
+		vim.keymap.set("n", "<leader>vws", function() vim.lsp.buf.workspace_symbol() end, opts)
+		vim.keymap.set("n", "<leader>vd", function() vim.diagnostic.open_float() end, opts)
+		vim.keymap.set("n", "[d", function() vim.diagnostic.goto_next() end, opts)
+		vim.keymap.set("n", "]d", function() vim.diagnostic.goto_prev() end, opts)
+		vim.keymap.set("n", "<leader>vca", function() vim.lsp.buf.code_action() end, opts)
+		vim.keymap.set("n", "<leader>vrr", function() vim.lsp.buf.references() end, opts)
+		vim.keymap.set("n", "<leader>vrn", function() vim.lsp.buf.rename() end, opts)
+		vim.keymap.set("i", "<C-h>", function() vim.lsp.buf.signature_help() end, opts)
+	end
+})
+
